Adds check to see if dpdk has been configured
If dpdk is configured then we do not want os-net-config to run multiple
times as it normally does during a deploy.  This is because when the NIC
is bound by dpdk, the kernel loses the interface and nic_mapping via
os-net-config will get screwed up.  With this fix we only run
os-net-config one time to setup the networks on compute nodes when dpdk
is used.